CSS replaced by Aphrodite CSS-in-JS for React components + components reorganized
