+++ assets/client/components/TitleBar.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { StyleSheet, css } from 'aphrodite';
 
-import { VitrineComponent } from '../VitrineComponent';
+import { VitrineComponent } from './VitrineComponent';
 
-import * as vitrineIcon from '../../images/vitrine.ico';
+import * as vitrineIcon from '../images/vitrine.ico';
 
 export class TitleBar extends VitrineComponent {
 	render() {
