+++ assets/client/sass/_contextMenus.scss
@@ -1,8 +1,8 @@
-@import 'bootstrapVariables';
+@import 'bootstrap.variables';
 
 .react-contextmenu {
 	min-width: 160px;
-	padding: 5px 0px;
+	padding: 5px 0;
 	margin: 2px 0 0;
 	font-size: 16px;
 	color: $text-color;
