+++ assets/client/sass/_reactDateTime.scss
@@ -1,0 +1,28 @@
+@import 'bootstrap.variables';
+
+.rdtPicker {
+	background-color: $body-bg;
+	border: none;
+
+	table tr td {
+		border-radius: $border-radius-base;
+	}
+	th {
+		border-bottom: 1px solid rgba(238, 238, 238, 0.15)
+	}
+	td.rdtDay:hover, td.rdtHour:hover, td.rdtMinute:hover,td.rdtSecond:hover, .rdtTimeToggle:hover {
+		background-color: #3E372E;
+		color: #AFACA7;
+		border-radius: $border-radius-base;
+	}
+	.rdtCounter .rdtBtn:hover ,td.rdtMonth:hover, td.rdtYear:hover button:hover, thead tr:first-child th:hover {
+		background: #3E372E;
+	}
+	td.rdtActive, td.rdtActive:hover {
+		background-color: $brand-primary;
+		color: #FFFFFF;
+	}
+	td.rdtToday:before {
+		border-bottom: 7px solid $brand-primary;
+	}
+}
