+++ package.json
@@ -62,7 +62,9 @@
     "jquery-contextmenu": "^2.6.2",
     "moment": "^2.18.1",
     "react": "^16.0.0",
+    "react-bootstrap": "^0.31.4",
     "react-contextmenu": "^2.8.0",
+    "react-datetime": "^2.10.3",
     "react-dom": "^16.0.0",
     "react-localization": "^0.1.2",
     "react-transition-group": "^2.2.1",
